Flatten the package hierarchy
