+++ collector/api/v1/api.go
@@ -5,7 +5,7 @@ import (
 
 	apexLog "github.com/apex/log"
 	"github.com/gin-gonic/gin"
-	"github.com/ooni/collector/collector/handler"
+	"github.com/ooni/collector/collector"
 	ginprometheus "github.com/zsais/go-gin-prometheus"
 )
 
@@ -16,7 +16,7 @@ var log = apexLog.WithFields(apexLog.Fields{
 
 // BindAPI bind all the request handlers and middleware
 func BindAPI(router *gin.Engine) error {
-	p := ginprometheus.NewPrometheus("oonicollector", handler.CustomMetrics)
+	p := ginprometheus.NewPrometheus("oonicollector", collector.CustomMetrics)
 	ignoredParams := []string{"reportID", "filename"}
 	p.ReqCntURLLabelMappingFn = func(c *gin.Context) string {
 		url := c.Request.URL.String()
@@ -33,15 +33,15 @@ func BindAPI(router *gin.Engine) error {
 	p.Use(router)
 
 	// This is to support legacy clients
-	router.POST("/report", handler.CreateReportHandler)
-	router.PUT("/report", handler.DeprecatedUpdateReportHandler)
-	router.POST("/report/:reportID", handler.UpdateReportHandler)
-	router.POST("/report/:reportID/close", handler.CloseReportHandler)
+	router.POST("/report", collector.CreateReportHandler)
+	router.PUT("/report", collector.DeprecatedUpdateReportHandler)
+	router.POST("/report/:reportID", collector.UpdateReportHandler)
+	router.POST("/report/:reportID/close", collector.CloseReportHandler)
 
 	v1 := router.Group("/api/v1")
-	v1.POST("/report", handler.CreateReportHandler)
-	v1.POST("/report/:reportID", handler.UpdateReportHandler)
-	v1.POST("/report/:reportID/close", handler.CloseReportHandler)
-	v1.POST("/measurement", handler.SubmitMeasurementHandler)
+	v1.POST("/report", collector.CreateReportHandler)
+	v1.POST("/report/:reportID", collector.UpdateReportHandler)
+	v1.POST("/report/:reportID/close", collector.CloseReportHandler)
+	v1.POST("/measurement", collector.SubmitMeasurementHandler)
 	return nil
 }
