+++ collector/errors.go
@@ -1,4 +1,4 @@
-package report
+package collector
 
 import "errors"
