+++ collector/handler.go
@@ -1,20 +1,13 @@
-package handler
+package collector
 
 import (
 	"errors"
 	"net/http"
 	"regexp"
 
-	apexLog "github.com/apex/log"
+	"github.com/apex/log"
 	"github.com/gin-gonic/gin"
-	"github.com/ooni/collector/collector/info"
-	"github.com/ooni/collector/collector/report"
 )
-
-var log = apexLog.WithFields(apexLog.Fields{
-	"pkg": "handler",
-	"cmd": "ooni-collector",
-})
 
 // CreateReportRequest what a client sends as a request to create a new report
 type CreateReportRequest struct {
@@ -45,7 +38,7 @@ func validateRequest(req *CreateReportRequest) error {
 
 // CreateReportHandler for report creation
 func CreateReportHandler(c *gin.Context) {
-	store := c.MustGet("Storage").(*report.Storage)
+	store := c.MustGet("Storage").(*Storage)
 
 	var req CreateReportRequest
 
@@ -58,7 +51,7 @@ func CreateReportHandler(c *gin.Context) {
 		return
 	}
 
-	reportID, err := report.CreateNewReport(store, req.TestName, req.ProbeASN, req.SoftwareName, req.SoftwareVersion)
+	reportID, err := CreateNewReport(store, req.TestName, req.ProbeASN, req.SoftwareName, req.SoftwareVersion)
 	if err != nil {
 		// XXX check this against the spec
 		c.JSON(http.StatusBadRequest, gin.H{
@@ -68,7 +61,7 @@ func CreateReportHandler(c *gin.Context) {
 	}
 
 	c.JSON(http.StatusOK, gin.H{
-		"backend_version":   info.Version,
+		"backend_version":   Version,
 		"report_id":         reportID,
 		"supported_formats": []string{"json"},
 	})
@@ -85,15 +78,15 @@ func DeprecatedUpdateReportHandler(c *gin.Context) {
 
 // UpdateReportRequest is used to update a report
 type UpdateReportRequest struct {
-	Content report.MeasurementEntry `json:"content" binding:"required"`
-	Format  string                  `json:"format"`
+	Content MeasurementEntry `json:"content" binding:"required"`
+	Format  string           `json:"format"`
 }
 
 // UpdateReportHandler appends to an open report
 func UpdateReportHandler(c *gin.Context) {
 	var err error
 
-	store := c.MustGet("Storage").(*report.Storage)
+	store := c.MustGet("Storage").(*Storage)
 	reportID := c.Param("reportID")
 
 	var req UpdateReportRequest
@@ -103,9 +96,9 @@ func UpdateReportHandler(c *gin.Context) {
 	}
 	entry := req.Content
 
-	measurementID, err := report.WriteEntry(store, reportID, &entry)
+	measurementID, err := WriteEntry(store, reportID, &entry)
 	if err != nil {
-		if err == report.ErrReportNotFound {
+		if err == ErrReportNotFound {
 			log.WithError(err).Debug("report not found error")
 			// XXX use the correct return value
 			c.JSON(http.StatusNotFound, gin.H{
@@ -129,10 +122,10 @@ func UpdateReportHandler(c *gin.Context) {
 
 // CloseReportHandler moves the report to the report-dir
 func CloseReportHandler(c *gin.Context) {
-	store := c.MustGet("Storage").(*report.Storage)
+	store := c.MustGet("Storage").(*Storage)
 	reportID := c.Param("reportID")
 
-	err := report.CloseReport(store, reportID)
+	err := CloseReport(store, reportID)
 	if err != nil {
 		// XXX return proper error
 		c.JSON(http.StatusNotAcceptable, gin.H{
@@ -149,9 +142,9 @@ func CloseReportHandler(c *gin.Context) {
 // SubmitMeasurementHandler is a handler for submitting a measurement in a
 // single request
 func SubmitMeasurementHandler(c *gin.Context) {
-	store := c.MustGet("Storage").(*report.Storage)
+	store := c.MustGet("Storage").(*Storage)
 	var (
-		entry    report.MeasurementEntry
+		entry    MeasurementEntry
 		reportID string
 	)
 
@@ -174,7 +167,7 @@ func SubmitMeasurementHandler(c *gin.Context) {
 		})
 	}
 	if reportID == "" {
-		rid, err := report.CreateNewReport(store, createReq.TestName,
+		rid, err := CreateNewReport(store, createReq.TestName,
 			createReq.ProbeASN, createReq.SoftwareName, createReq.SoftwareVersion)
 		if err != nil {
 			c.JSON(http.StatusBadRequest, gin.H{
@@ -183,14 +176,14 @@ func SubmitMeasurementHandler(c *gin.Context) {
 		}
 		reportID = rid
 	}
-	measurementID, err := report.WriteEntry(store, reportID, &entry)
+	measurementID, err := WriteEntry(store, reportID, &entry)
 	if err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{
 			"error": err.Error(),
 		})
 	}
 	if shouldClose == true {
-		report.CloseReport(store, reportID)
+		CloseReport(store, reportID)
 	}
 	c.JSON(http.StatusOK, gin.H{
 		"report_id":      reportID,
