+++ collector/info.go
@@ -1,4 +1,4 @@
-package info
+package collector
 
 import "fmt"
