+++ collector/storage.go
@@ -1,4 +1,4 @@
-package report
+package collector
 
 import (
 	"context"
